+++ src/elite/test/EliteSharpLib.Fakes/FakeView.cs
@@ -1,0 +1,26 @@
+// 'Elite - The Sharp Kind' - Andy Hawkins 2023.
+// 'Elite - The New Kind' - C.J.Pinder 1999-2001.
+// Elite (C) I.Bell & D.Braben 1984.
+
+using EliteSharpLib.Views;
+
+namespace EliteSharpLib.Fakes;
+
+internal sealed class FakeView : IView
+{
+    public void Draw()
+    {
+    }
+
+    public void HandleInput()
+    {
+    }
+
+    public void Reset()
+    {
+    }
+
+    public void Update()
+    {
+    }
+}